// portfolio/sanity.types.ts
export type PROFILE_QUERYResult =
  | {
      email: null;
      phone: null;
      location: null;
      socialLinks: null;
    }
  | {
      email: null;
      phone: null;
      location: string | null;
      socialLinks: null;
    }
  | {
      email: string | null;
      phone: null;
      location: null;
      socialLinks: null;
    }
  | {
      email: string | null;
      phone: string | null;
      location: string | null;
      socialLinks: {
        github?: string;
        linkedin?: string;
        twitter?: string;
        website?: string;
        medium?: string;
        devto?: string;
        youtube?: string;
        stackoverflow?: string;
      } | null;
    }
  | null;